Fix character loading
I think I see what you are trying to do here, and this was the quickest
way to get things back on track. However, it does not accomplish what I
think you are actually going for. We should talk about what you are
doing and you can make a plan from there.

- Change JSON file to js
- File organization change
- Remove unusable json loader (via `npm remove -D json-loader`)