De-globalize character var usage
Technically `character` is not a "global", but it is in the root scope
of this module and was being used as though it was always available
there. As these components are moved elsewhere, and more importantly, as
character becomes a server-fectched value, we cannot depend on this
availability.

So I have adjusted so that the root-level `character` is only used in
one place (LoadApp). Everything else gets `character` from an argument,
context, or props.