Add recoil to `dependencies`
And remove from `devDependencies`. This is often a bit of a theoretical
distinction, especially if you use a build tool that packages your
application when you deploy. However, since recoil will ship with your
code (just like react will ship with your code), it is best for it to be
in the "deps" list.