Refactor skill forms
Use one form for both update and create.

If we pass along the desired actions to take on form submission, we can
keep the saving logic separate from the display logic. That way, the
parent can determine what should happen when the form is submitted and
the child can just blindly display data.

Another added benefit to doing it this way is that we can use the same
function for the actual persistence of the data, whereas before we had
to define the same thing (and many other things) twice.

The one behavior change here is that you can now edit the name of old
skills. If that is undesired, we can prevent it directly.