Add type info to several key files
Types in typescript have ZERO effect on the final javascript that runs
in the browser. They are only there as hints for the compiler so that it
can make you aware of places where you are using your variables
incorrectly.

The running code in the browser before this commit is *ecactly the same*
as the running code after this commit.

Any errors that Typescript has found in the files that are modified by
this commit (except for server/index.ts) are *actual* errors that you
should address in some way.